Dont execute nasm when compiler exits with error

+++ testes/compilador.sh
@@ -4,7 +4,10 @@ SRC=$2
 echo "$COMPILADOR $SRC"
 eval "$COMPILADOR $SRC"
 
-echo "nasm -felf64 ${SRC}.s && gcc ${SRC}.o"
-nasm -felf64 ${SRC}.s && gcc ${SRC}.o
-rm ${SRC}.o
+if [ $? -eq 0 ]
+then
+	echo "nasm -felf64 ${SRC}.s && gcc ${SRC}.o"
+	nasm -felf64 ${SRC}.s && gcc ${SRC}.o
+	rm ${SRC}.o
+fi
